对执行结果进行优化

+++ src/screen_shot.py
@@ -1,4 +1,5 @@
 import datetime
+import os
 
 
 class screen_shot():
@@ -10,15 +11,26 @@ class screen_shot():
 
     def screen_shot(self, adb, device, save_path, delete_shot):
         file_name = self.get_file_name()
-        # 截图
         screen_cap = adb.run("-s " + device + " shell screencap -p /sdcard/" + file_name)
-        print(screen_cap)
-        adb.run("-s " + device + " pull /sdcard/" + file_name + " " + save_path)
+        if self.check_result(screen_cap):
+            # 截图失败
+            print('截图失败：' + screen_cap)
+            return
+        pull = adb.run("-s " + device + " pull /sdcard/" + file_name + " " + save_path)
+        if self.check_result(pull):
+            # 拷贝出错了
+            print('拷贝截图失败：' + pull)
+            return
+        print("截图成功，文件已保存至：" + save_path + file_name)
         # 删除截图
         if delete_shot:
             adb.run("shell rm /sdcard/" + file_name)
-        print("截图成功文件已保存至：" + save_path + file_name)
         pass
 
     def get_file_name(self):
         return datetime.datetime.now().strftime('%Y-%m-%d-%H-%M-%S') + '.png'
+
+    # 分析返回结果
+    def check_result(self, result):
+        if result.startswith('adb: error'):
+            return True
